feat: implement `transformSources()` w/ domain model (#356)
Introduces a domain model for _Package_ and _Entry Point_. Serves as a foundation for a set of interesting ideas: what if... we could analyze an entry points `import` statements and dynamically create a minimal rollup configuration? What if... we further can detect that an entry point depends on another entry point (limitation: within same package) so we could create a build graph and auto-infer the tsconfig?

BREAKING CHANGES:
Introduces a domain model for _Package_ and _Entry Point_ (as defined in Angular Package Format Specification Glossary). Refactors the source code transformation pipeline. Albeit the refactoring was undertaken with care and integration tests were not changed, it may cause undesired behaviour.