feat: update `lib` compiler option to `es2018` (#956)
BREAKING CHANGE:

The following `peerDependencies` are now required
- `tslib: ^1.9.0`
- `typescript: ^2.7.0`

The default for `lib.languageLevel`  is now `es2018`. In case this needs to be changed. This can to be done by overriding the tsconfig. More on how to override the tsconfig can be found here: https://github.com/dherges/ng-packagr/blob/master/docs/override-tsconfig.md